useless/requirements.txt
flask
gunicorn
psycopg2-binary
pandas
openpyxl
python-docx
python-dotenv
